requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
openai==1.58.1
chromadb==0.6.3
python-docx==1.1.2
PyMuPDF==1.25.2
